Update before_after_operation_rule.md
I made a few quick edits that represent where code is executed, even if the best practices of linking in external scripts is used.

The diagram doesn't show the IQService component. It only shows the VA. VA reaches out to IDN for a task. VA reaches out to IQService (some logging is done there). A process is started up to run the rule which can included.

It's unclear to me that all the Rule Types have the same inputs and outputs. E.g. Does a BeforeModify Rule have access to manipulate the provisioning result? In looking through rules I have access to there seem to be things passed as arguments, as environment variables, and possibly via files. I can't speak to what current best practices are.

Ideally if the Rule Types have different inputs/outputs this would be called out for each.

All logging examples I've seen are to local files on the IQService servers. This document refers to logs being in the ccg log and being forwarded to SailPoint. Given these execute on the IQService server I suggest what gets logged and where it gets log be validated and reflected on this site.

Some information on error handling is would help. For example, it is appears possible to manipulate the provisioning response in an AfterCreate rule. This would get back to IDN (or IIQ) for and may be visible in the UI and error messages.

Documentation on the SailPoint objects contained in the Utils.dll would be beneficial. Something like the JavaDoc for beanshell rules.

I don't know if the PowerShell rules are used as often as the beanshell rules, such a reference could help.